+++ src/randomizer.js
@@ -1,6 +1,18 @@
-import React from 'react';
+const randomIndex = (array) => {
+   const index = Math.floor(Math.random() * array.length);
+   return array[index]
+}
 
+const randomThree = (array) => {
+   let one = randomIndex(array);
+   let two = randomIndex(array);
+   let three = randomIndex(array);
+   if (one === two || one === three) { one = randomIndex(array) }
+   if (two === one || two === three) { two = randomIndex(array) }
+   if (three === two || three === one) { three = randomIndex(array) }
+   const newArray = []
+   newArray.push(one, two, three)
+   return newArray
+}
 
-
-
-export default Randomizer;
+export default randomThree;
